requirements.txt
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bitsandbytes==0.39.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
datasets==2.13.1
dill==0.3.6
einops==0.6.1
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gradio==3.32.0
gradio-client==0.2.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
importlib-resources==5.12.0
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.183
linkify-it-py==2.0.2
lit==16.0.5
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openapi-schema-pydantic==1.2.4
orjson==3.8.14
packaging==23.1
pandas==2.0.1
peft==0.3.0
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
PyMuPDF==1.22.5
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing-extensions==4.6.2
typing-inspect==0.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.2
uvicorn==0.22.0
websockets==11.0.3
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
